fixed request logger
